+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Handling of naming for more modules of one type (CPX-E and -AP)
 - Added black job to CI/CD
 - Added tests for other CPX-AP modules
 - Added vabx module for vtux access